Added playback mode control to animatlab. You can now specify that you want the playback time step to match the physics step, or a preset value. Also added timing values for the step, physics, and neural that can be charted. Added a test for this.
Got rid of some unused items in the GUI.

Fixed a timing issue with the HUD. It no longer draws now if no window is open and it only updates every 0.1 second.

+++ Libraries/IntegrateFireSim/Neuron.cpp
@@ -841,7 +841,7 @@ void Neuron::CalculateFiringFreq(IntegrateFireNeuralModule *lpNS)
 {
 	if(m_bSpike)
 	{
-		double dblDiff = (lpNS->GetCurrentTime() - m_fltLastSpikeTime)/1000;
+		double dblDiff = (lpNS->GetCurrentTime() - m_fltLastSpikeTime)/(double) 1000.0;
 		
 		if(dblDiff > 0)
 			m_fltFiringFreq = 1/dblDiff;
